change get_mode() to an ext-function to avoid loop-call errors
