+++ CTFd/plugins/ctfd-owl/control_utils.py
@@ -8,15 +8,6 @@ from flask import session
 
 class ControlUtil:
     @staticmethod
-    def get_mode():
-        mode = utils.get_config("user_mode")
-        if mode == "teams":
-            user_id = current_user.get_current_user().team_id
-        else:
-            user_id = current_user.get_current_user().id
-        return user_id
-
-    @staticmethod
     def new_container(user_id, challenge_id):
         rq = DockerUtils.up_docker_compose(user_id=user_id, challenge_id=challenge_id)
         if rq != False:
